Crate ps2

Enums

KeyboardType

Functions

check_mouse_id

check the mouse’s id

command_to_keyboard

write command to the keyboard and return the result

command_to_mouse

write command to the mouse and return the result

data_to_port1

write data to the first ps2 data port and return the response

data_to_port2

write data to the second ps2 data port and return the response

handle_mouse_packet

read mouse data packet

init_ps2_port1

initialize the first ps2 data port

init_ps2_port2

initialize the second ps2 data port

keyboard_detect

detect the keyboard’s type

keyboard_led

set LED status of the keyboard parameter : 0: ScrollLock; 1: NumberLock; 2: CapsLock

keyboard_scancode_set

set the scancode set of the keyboard 0: get the current set; 1: set 1 2: set 2; 3: set 3

mouse_packet_streaming

enable or disable the packet streaming also return the vec of data previously in the buffer which may be useful

mouse_resend

resend the most recent packet again

mouse_resolution

set the resolution of the mouse parameter : 0x00: 1 count/mm; 0x01 2 count/mm; 0x02 4 count/mm; 0x03 8 count/mm;

ps2_clean_buffer

clean the PS2 data port (0x60) output buffer also return the vec of data previously in the buffer which may be useful

ps2_read_config

read the config of the ps2 port

ps2_read_data

read dat from ps2 data port (0x60)

ps2_status_register

read the ps2 status register

ps2_write_command

write command to the command ps2 port (0x64)

ps2_write_config

write the new config to the ps2 command port (0x64)

reset_mouse

reset the mouse

set_mouse_id

set the mouse ID (3 or 4 ) by magic sequence 3 means that the mouse has scroll 4 means that the mouse has scroll, fourth and fifth buttons

set_sampling_rate

set ps2 mouse’s sampling rate

test_ps2_port1

test the first ps2 data port

test_ps2_port2

test the second ps2 data port

write_to_second_output_buffer

write data to the second ps2 output buffer