Crate ps2

Driver for the PS/2 interface, which supports legacy keyboard and mouse devices.

The init() function sets up the single system-wide PS2Controller, which can then be used to obtain a reference to a PS2Keyboard or PS2Mouse.

Structs

ControllerConfigurationByte

Configuration values used to enable or disable features of the PS/2 controller.

ControllerToHostStatus

LEDState

MousePacket3

MousePacket4

MousePacketGeneric

PS2Controller

An initialized PS/2 controller that can be used to communicate with legacy PS/2 keyboard and mouse devices.

PS2Keyboard

A reference to the PS/2 Keyboard connected to the PS/2 Controller.

PS2Mouse

A reference to the PS/2 Mouse connected to the PS/2 Controller.

Enums

DeviceToHostResponse

HostToControllerCommand

Commands that can be sent to the PS/2 Controller.

HostToKeyboardCommand

HostToKeyboardCommandOrData

HostToMouseCommand

HostToMouseCommandOrData

KeyboardType

MouseId

See MousePacket and its enum variants for further information.

MousePacket

MouseResolution

MouseSampleRate

PortTestResult

ScancodeSet

Functions

init

Initializes the PS/2 controller, and the first and second PS/2 ports (if they exist).